backend/analysis.py

def calculate_feature_to_fix_ratio(commit_history):
    """
    Placeholder for calculating feature-to-fix ratio.
    Needs actual commit message parsing to identify feature vs. fix commits.
    """
    # Dummy implementation
    features = 0
    fixes = 0
    for commit in commit_history:
        if "feat" in commit.lower(): # Simplified check
            features += 1
        elif "fix" in commit.lower(): # Simplified check
            fixes += 1
    return features / (fixes if fixes > 0 else 1)

def calculate_code_churn_volatility(commit_history):
    """
    Placeholder for calculating code churn volatility.
    Needs access to diffs for each commit to determine lines added/deleted.
    """
    # Dummy implementation
    return len(commit_history) * 0.1 # Placeholder value

def calculate_commit_cadence(commit_history):
    """
    Placeholder for calculating commit cadence (rhythm).
    Needs commit timestamps to determine frequency and patterns.
    """
    # Dummy implementation
    return len(commit_history) / 30.0 # Placeholder: commits per month

def analyze_glyph_health(repo_path):
    """
    Main function to analyze the health of a Glyph based on its repository.
    This will eventually interact with GitPython to get actual commit data.
    """
    from git import Repo, exc
    
    commit_history = []
    try:
        repo = Repo(repo_path)
        for commit in repo.iter_commits():
            commit_history.append(commit.message)
    except exc.InvalidGitRepositoryError:
        print(f"Error: {repo_path} is not a valid Git repository.")
        return {
            "feature_to_fix_ratio": 0.0,
            "code_churn_volatility": 0.0,
            "commit_cadence": 0.0,
            "stability_graph_data": [],
            "development_tempo_data": []
        }

    feature_to_fix = calculate_feature_to_fix_ratio(commit_history)
    churn_volatility = calculate_code_churn_volatility(commit_history)
    commit_rhythm = calculate_commit_cadence(commit_history)

    return {
        "feature_to_fix_ratio": feature_to_fix,
        "code_churn_volatility": churn_volatility,
        "commit_cadence": commit_rhythm,
        "stability_graph_data": [10, 12, 15, 13, 16, 14], # Dummy data for a graph
        "development_tempo_data": [5, 7, 6, 8, 7, 9] # Dummy data for a chart
    }

if __name__ == "__main__":
    # Example usage
    repo_path = "/path/to/your/repo" # Replace with actual repo path
    health_metrics = analyze_glyph_health(repo_path)
    print("Glyph Health Metrics:")
    for key, value in health_metrics.items():
        print(f"- {key}: {value}")
