backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
GitPython==3.1.43
svgwrite==1.4.1
python-dotenv==1.0.1
starlette==0.37.2
itsdangerous==2.2.0
python-multipart==0.0.9
PyYAML==6.0.1
nltk==3.8.1
cairosvg==2.7.1
authlib==1.3.0